pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dxt-core"
version = "0.0.1"
description = "A lightweight Extract-Load tool for building simple, reliable batch data transfer pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DXT Contributors"}
]
keywords = ["etl", "elt", "data", "pipeline", "extract", "load", "yaml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pyarrow>=14.0.0",
    "rich>=13.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
postgres = [
    "psycopg[binary]>=3.1.0",
]
all = [
    "dxt[postgres]",
]

[project.scripts]
dxt = "dxt.cli:app"

[project.urls]
Homepage = "https://github.com/Untitled-Research/dxt/dxt"
Documentation = "https://github.com/Untitled-Research/dxt/dxt#readme"
Repository = "https://github.com/Untitled-Research/dxt/dxt"
Issues = "https://github.com/Untitled-Research/dxt/dxt/issues"

[tool.hatch.build.targets.wheel]
packages = ["dxt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=dxt --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
